# optimize code to reduce execution time