--- a/pkg/training/tensorboard.go
+++ b/pkg/training/tensorboard.go
@@ -1,39 +1,31 @@
 package training
 
 import (
-	"context"
 	"fmt"
 
 	"github.com/kubeflow/arena/pkg/apis/config"
-	"github.com/kubeflow/arena/pkg/arenacache"
 	log "github.com/sirupsen/logrus"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-func tensorboardURL(name, namespace string) (url string, err error) {
+func tensorboardURL(name, namespace string, services []*v1.Service, nodes []*v1.Node) (url string, err error) {
 
 	var (
 		port int32
 	)
-	clientset := config.GetArenaConfiger().GetClientSet()
-	// 1. Get port
-	labels := map[string]string{
-		"release": name,
-		"role":    "tensorboard",
+	var service *v1.Service
+	for _, svc := range services {
+		if svc.Labels["release"] == name && namespace == svc.Namespace {
+			service = svc
+			break
+		}
 	}
-	serviceList, err := listServices(clientset, namespace, labels)
-	if err != nil {
-		return "", err
-	}
-
-	if len(serviceList.Items) == 0 {
+	if service == nil {
 		log.Debugf("Failed to find the tensorboard service due to service"+
 			"List is empty when selector is release=%s,role=tensorboard.", name)
 		return "", nil
 	}
-
-	service := serviceList.Items[0]
 	portList := service.Spec.Ports
 	if len(portList) == 0 {
 		log.Debugf("Failed to find the tensorboard service due to ports list is empty.")
@@ -52,27 +44,19 @@ func tensorboardURL(name, namespace string) (url string, err error) {
 	port = portList[0].NodePort
 
 	// 2. Get address
-	nodeList := &v1.NodeList{}
-	if config.GetArenaConfiger().IsDaemonMode() {
-		err = arenacache.GetCacheClient().List(context.Background(), nodeList)
-	} else {
-		nodeList, err = clientset.CoreV1().Nodes().List(metav1.ListOptions{})
-	}
-	node := v1.Node{}
-	findReadyNode := false
+	var node *v1.Node
 
-	for _, item := range nodeList.Items {
-		if isNodeReady(item) {
-			node = item
-			findReadyNode = true
+	for _, n := range nodes {
+		if isNodeReady(*n) {
+			node = n
 			break
 		}
 	}
 
-	if !findReadyNode {
+	if node == nil {
 		return "", fmt.Errorf("Failed to find the ready node for exporting tensorboard.")
 	}
-	url = fmt.Sprintf("http://%s:%d", getNodeInternalAddress(node), port)
+	url = fmt.Sprintf("http://%s:%d", getNodeInternalAddress(*node), port)
 
 	return url, nil
 }
@@ -97,4 +81,36 @@ func getNodeInternalAddress(node v1.Node) string {
 		}
 	}
 	return address
+}
+
+func PrepareServicesAndNodesForTensorboard(jobs []TrainingJob, allNamespaces bool) ([]*v1.Service, []*v1.Node) {
+	clientset := config.GetArenaConfiger().GetClientSet()
+	services := []*v1.Service{}
+	nodes := []*v1.Node{}
+	var err error
+	if len(jobs) == 0 {
+		return services, nodes
+	}
+	labels := map[string]string{
+		"release": "",
+		"role":    "tensorboard",
+	}
+	if len(jobs) == 1 {
+		labels["release"] = jobs[0].Name()
+	}
+	namespace := jobs[0].Namespace()
+	if allNamespaces {
+		namespace = metav1.NamespaceAll
+	}
+	services, err = listServices(clientset, namespace, labels)
+	if err != nil {
+		log.Errorf("failed to list k8s services when query dashboard url,reason: %v", err)
+		services = []*v1.Service{}
+	}
+	nodes, err = listNodes(clientset, map[string]string{})
+	if err != nil {
+		log.Errorf("failed to list nodes when query dashboard url,reason: %v", err)
+		nodes = []*v1.Node{}
+	}
+	return services, nodes
 }